requirements.txt
Cython
scikit-image
tensorflow
keras
requests
Pillow
h5py
tensorboard_logger
